Test ability to `ssh` into specified host
This adds a new step that attempts to `ssh` into the host before delegating
the build.  This matters most when delegating builds as the root user (such as
in a multi-user Nix installation) because the root user might not have the
destination machine added as a known host.  If the destination is not a known
host then the distributed build support will silently fail, which this change
fixes.